changes to move the wrong port exeption to the end of the namespace - otherwise
the design form fails - a microsoft requirement.
Also removed the while loop in tmrstepperrequests to try to avoid the hang which sometimes happens if no data is received.  The timer interval will need to be increased though if the hang is a problem, as the high timer frequency has the same effect as the while loop !